python_grpc/requirements.txt

grpcio==1.33.2
grpcio-tools==1.33.2
pymongo==3.11.0
redis==3.5.3
requests==2.24.0